00-bootcamp-project/pyproject.toml
[tool.poetry]
name = "00-bootcamp-project"
version = "0.1.0"
description = ""
authors = ["Krai Chamnivikaipong <krai020913@gmail.com>"]

[tool.poetry.dependencies]
python = "3.10.13"
requests = "2.29.0"
pysftp = "0.2.9"
psycopg2 = "2.9.6"
google-cloud-bigquery = "^3.17.2"
google-cloud-storage = "^2.14.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"